reorganised and added a pixel heart for likes
